Refactor i386 layout
