+++ kernel/include/Log.h
@@ -28,7 +28,7 @@
 /***************************************************************************/
 
 #include "Base.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 #include "Task.h"
 
 /***************************************************************************/
