+++ kernel/include/Memory.h
@@ -40,7 +40,7 @@
 /************************************************************************/
 
 #include "Base.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 
 /************************************************************************/
