+++ kernel/include/Process.h
@@ -29,7 +29,7 @@
 
 #include "Base.h"
 #include "Driver.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 #include "ID.h"
 #include "List.h"
 #include "Mutex.h"
