+++ kernel/include/Task.h
@@ -32,7 +32,7 @@
 /************************************************************************/
 
 #include "Base.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 #include "List.h"
 #include "Mutex.h"
 #include "User.h"
