+++ kernel/source/Clock.c
@@ -24,7 +24,7 @@
 
 #include "../include/Clock.h"
 
-#include "../include/I386.h"
+#include "../include/arch/i386/I386.h"
 #include "../include/InterruptController.h"
 #include "../include/Kernel.h"
 #include "../include/Log.h"
