+++ kernel/source/Fault.c
@@ -24,7 +24,7 @@
 
 #include "../include/Console.h"
 #include "../include/Heap.h"
-#include "../include/I386.h"
+#include "../include/arch/i386/I386.h"
 #include "../include/Kernel.h"
 #include "../include/Log.h"
 #include "../include/Memory.h"
