+++ kernel/source/InterruptFrame.c
@@ -146,7 +146,7 @@
 #include "../include/InterruptFrame.h"
 
 #include "../include/Base.h"
-#include "../include/I386.h"
+#include "../include/arch/i386/I386.h"
 #include "../include/Log.h"
 #include "../include/Memory.h"
 #include "../include/String.h"
