+++ kernel/source/autotest/Autotest-I386-MCI.c
@@ -25,7 +25,7 @@
 #include "../../include/Autotest.h"
 #include "../../include/Base.h"
 #include "../../include/Heap.h"
-#include "../../include/I386-MCI.h"
+#include "../../include/arch/i386/I386-MCI.h"
 #include "../../include/Log.h"
 #include "../../include/String.h"
