+++ kernel/source/drivers/Keyboard.c
@@ -26,7 +26,7 @@
 
 #include "Base.h"
 #include "Console.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 #include "InterruptController.h"
 #include "Kernel.h"
 #include "Log.h"
