+++ kernel/source/drivers/VESA.c
@@ -23,7 +23,7 @@
 \************************************************************************/
 
 #include "GFX.h"
-#include "I386.h"
+#include "arch/i386/I386.h"
 #include "Kernel.h"
 #include "Log.h"
